\name{affyPlatforms}
\alias{affyPlatforms}
\title{Available Affymetrix platforms for SNP arrays}
\description{
  Provides a listing of available Affymetrix platforms currently
  supported by the R package oligo
}
\usage{
affyPlatforms()
}


\value{
  A vector of class character.
}
\author{R. Scharpf}
\examples{
affyPlatforms()
}
\keyword{list}
\keyword{misc}

