\name{assayDataList-methods}
\docType{methods}
\alias{assayDataList}
\alias{assayDataList-methods}

\alias{assayDataList,oligoSetList-method}
\title{Accessor for slot assayDataList in Package \pkg{oligoClasses} }
\description{
  Accessor for slot assayDataList in Package \pkg{oligoClasses}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "gSetList")}}{
  An object inheriting from class \code{gSetList}.
}

\item{\code{signature(object = "oligoSetList")}}{
  An object inheriting from class \code{gSetList}.
}
}}
\keyword{methods}
\keyword{manip}

