\name{Deprecated}
\title{oligoClasses Deprecated}
\alias{featuresInRange}
\alias{featuresInRange,SnpSet2,RangedDataCNV-method}
\alias{RangedDataCNV-class}
\alias{RangedDataHMM-class}
\alias{RangedDataCNV}
\alias{RangedDataHMM}
\alias{RangedDataCBS}
\alias{RangedDataCBS-class}
\alias{RangedDataCopyNumber-class}
\alias{state,RangedDataCNV-method}
%\alias{coverage2}
\alias{coverage2,RangedDataCNV-method}
\alias{mean,RangedDataCBS-method}
\alias{sampleNames,RangedDataCNV-method}
\alias{chromosome,RangedDataCNV-method}
\alias{findOverlaps,RangedDataCNV,SnpSet-method}
\alias{findOverlaps,RangedDataCNV,CNSet-method}
\alias{findOverlaps,RangedDataCNV,RangedDataCNV-method}
\alias{findOverlaps,RangedDataHMM,RangedDataHMM-method}
\alias{findOverlaps,RangedDataCNV,AnnotatedDataFrame-method}
\alias{findOverlaps,AnnotatedDataFrame,RangedDataCNV-method}
\alias{sampleNames<-,RangedDataCNV,character-method}
\description{ The function, class, or data object you asked for has been
  deprecated.}
\keyword{internal}
