\name{featureDataList-methods}
\docType{methods}
\alias{featureDataList}
\alias{featureDataList-methods}
\title{ Accessor for slot \code{featureDataList} in Package \pkg{oligoClasses} ~~}
\description{
 Accessor for slot \code{featureDataList} in Package \pkg{oligoClasses} ~~
}
\section{Methods}{
\describe{

\item{\code{signature(object = "gSetList")}}{
  An object inheriting from class \code{gSetList}.
}
}}
\keyword{methods}
\keyword{manip}
