\name{ff_or_matrix-class}
\Rdversion{1.1}
\docType{class}
\alias{ff_or_matrix-class}

\title{Class \code{"ff_or_matrix"}}
\description{
  A class union of 'ffdf', 'ff_matrix', and 'matrix'
}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{GenomeAnnotatedDataFrameFrom}{\code{signature(object = "ff_or_matrix")}: ... }
	 }
}

\author{
R. Scharpf
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  \code{\link[ff]{ff}},   \code{\link[ff]{ffdf}}
}
\examples{
showClass("ff_or_matrix")
}
\keyword{classes}
