\name{ffdf-class}
\Rdversion{1.1}
\docType{class}
\alias{ffdf-class}
\alias{list_or_ffdf-class}


\title{Class "ffdf"}

\description{

	Extended package ff's class definitions for ff to S4.	

}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{ffdf} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
Class \code{"\linkS4class{list_or_ffdf}"}, directly.
}
\section{Methods}{
No methods defined with class "ffdf" in the signature.
}
\keyword{classes}
