\name{generics}
\alias{baf}
\alias{lrr}

\title{ Miscellaneous generics.  Methods defined in packages that depend
  on oligoClasses}
\description{Miscellaneous generics.  Methods defined in packages that depend
  on oligoClasses
}
\usage{
baf(object)
lrr(object)
}
\arguments{
  \item{object}{A \code{eSet}-derived class.}
}
\author{
R. Scharpf
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}
