\name{getBar}
\alias{getBar}

\title{
  Gets a bar of a given length.
}
\description{
  Gets a bar of a given length.
}
\usage{
getBar(width = getOption("width"))
}

\arguments{
  \item{width}{desired length of the bar.}
}

\value{character string.}

\author{
  Benilton S Carvalho
}

\examples{
message(getBar())
}

\keyword{manip}
