\name{length-methods}
\docType{methods}
\alias{length-methods}
\alias{length,FeatureSet-method}
\title{Number of samples for FeatureSet-like objects.}
\description{
  Number of samples for FeatureSet-like objects.
}
\section{Methods}{
\describe{
\item{x = "FeatureSet"}{Number of samples}
}}
\keyword{methods}
