\name{library2}
\alias{library2}
\title{
  Supress package startup messages when loading a library
}
\description{
  Supress package startup messages when loading a library
}
\usage{
library2(...)
}
\arguments{
  \item{\dots}{ arguments to \code{library}
}
}


\author{
R. Scharpf
}

\seealso{
\code{\link{library}}
}
\examples{
library2("Biobase")
}
\keyword{manip}
