\name{parStatus}
\alias{parStatus}

\title{
  Checks if oligo/crlmm can use parallel resources.
}

\description{
  Checks if oligo/crlmm can use parallel resources (needs ff and snow
  package, in addition to options(cluster=makeCluster(...)).
}
\usage{
parStatus()
}
\value{logical}

\author{
  Benilton S Carvalho
}

\keyword{manip}

