\name{pmFragmentLength-methods}
\docType{methods}
\alias{pmFragmentLength}
\alias{pmFragmentLength-methods}
\alias{pmFragmentLength,AffySNPPDInfo-method}
\title{Information on Fragment Length}
\description{
  This method will return the fragment length for PM probes.
}
\section{Methods}{
\describe{
\item{object = "AffySNPPDInfo"}{On \code{AffySNPPDInfo} objects, it will
return the fragment length that contains the SNP in question.}
}}
\keyword{methods}

