\name{sampleNames-methods}
\docType{methods}
\alias{sampleNames-methods}
\alias{sampleNames,FeatureSet-method}

\title{Sample names for FeatureSet-like objects}
\description{
  Returns sample names for FeatureSet-like objects.
}
\section{Methods}{
  \describe{
    \item{object = "FeatureSet"}{Sample names}
  }
  }

\keyword{methods}

